Separate rule engine service and rule action builder
