+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/programrule/DefaultProgramRuleService.java
@@ -27,14 +27,11 @@
  */
 package org.hisp.dhis.tracker.programrule;
 
-import static org.hisp.dhis.tracker.validation.validator.ValidationUtils.needsToValidateDataValues;
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -42,21 +39,10 @@ import java.util.stream.Stream;
 import lombok.RequiredArgsConstructor;
 
 import org.apache.commons.collections4.ListUtils;
-import org.apache.commons.lang3.StringUtils;
-import org.hisp.dhis.common.BaseIdentifiableObject;
 import org.hisp.dhis.program.Program;
 import org.hisp.dhis.program.ProgramInstance;
-import org.hisp.dhis.program.ProgramStage;
 import org.hisp.dhis.program.ProgramStageInstance;
 import org.hisp.dhis.programrule.engine.ProgramRuleEngine;
-import org.hisp.dhis.rules.models.RuleAction;
-import org.hisp.dhis.rules.models.RuleActionAssign;
-import org.hisp.dhis.rules.models.RuleActionError;
-import org.hisp.dhis.rules.models.RuleActionErrorOnCompletion;
-import org.hisp.dhis.rules.models.RuleActionSetMandatoryField;
-import org.hisp.dhis.rules.models.RuleActionShowError;
-import org.hisp.dhis.rules.models.RuleActionShowWarning;
-import org.hisp.dhis.rules.models.RuleActionWarningOnCompletion;
 import org.hisp.dhis.rules.models.RuleEffects;
 import org.hisp.dhis.trackedentity.TrackedEntityInstance;
 import org.hisp.dhis.trackedentityattributevalue.TrackedEntityAttributeValue;
@@ -64,26 +50,12 @@ import org.hisp.dhis.tracker.bundle.TrackerBundle;
 import org.hisp.dhis.tracker.converter.RuleEngineConverterService;
 import org.hisp.dhis.tracker.converter.TrackerConverterService;
 import org.hisp.dhis.tracker.domain.Attribute;
-import org.hisp.dhis.tracker.domain.DataValue;
 import org.hisp.dhis.tracker.domain.Enrollment;
 import org.hisp.dhis.tracker.domain.Event;
-import org.hisp.dhis.tracker.domain.TrackedEntity;
-import org.hisp.dhis.tracker.domain.TrackerDto;
 import org.hisp.dhis.tracker.preheat.TrackerPreheat;
-import org.hisp.dhis.tracker.programrule.implementers.RuleActionType;
-import org.hisp.dhis.tracker.programrule.implementers.enrollment.ActionRule;
-import org.hisp.dhis.tracker.programrule.implementers.enrollment.AssignActionRule;
-import org.hisp.dhis.tracker.programrule.implementers.enrollment.ErrorActionRule;
-import org.hisp.dhis.tracker.programrule.implementers.enrollment.ErrorOnCompleteActionRule;
-import org.hisp.dhis.tracker.programrule.implementers.enrollment.MandatoryActionRule;
-import org.hisp.dhis.tracker.programrule.implementers.enrollment.SyntaxErrorActionRule;
-import org.hisp.dhis.tracker.programrule.implementers.enrollment.WarningActionRule;
-import org.hisp.dhis.tracker.programrule.implementers.enrollment.WarningOnCompleteActionRule;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import com.google.common.collect.Lists;
 
 /**
  * @author Enrico Colasante
@@ -102,6 +74,8 @@ class DefaultProgramRuleService
 
     private final TrackerConverterService<Attribute, TrackedEntityAttributeValue> attributeValueTrackerConverterService;
 
+    private final ActionRuleBuilder actionRuleBuilder;
+
     @Override
     @Transactional( readOnly = true )
     public void calculateRuleEffects( TrackerBundle bundle, TrackerPreheat preheat )
@@ -112,175 +86,12 @@ class DefaultProgramRuleService
                 calculateProgramEventRuleEffects( bundle, preheat ),
                 calculateTrackerEventRuleEffects( bundle, preheat ) ) );
 
-        // The following converting/filtering process should be responsibility of a different class
-
         // This is needed for bunlde side effects process
         bundle.setRuleEffects( ruleEffects );
 
         // These are needed for rule engine validation
-        bundle.setEnrollmentActionRules( filterEnrollments( ruleEffects, bundle ) );
-        bundle.setEventActionRules( filterEvents( ruleEffects, bundle ) );
-    }
-
-    private Map<TrackerDto, List<ActionRule>> filterEnrollments( List<RuleEffects> ruleEffects, TrackerBundle bundle )
-    {
-        return ruleEffects.stream()
-            .filter( RuleEffects::isEnrollment )
-            .filter( e -> bundle.findEnrollmentByUid( e.getTrackerObjectUid() ).isPresent() )
-            .collect( Collectors.toMap( e -> bundle.findEnrollmentByUid( e.getTrackerObjectUid() ).get(),
-                e -> buildEnrollmentActionRules( bundle.findEnrollmentByUid( e.getTrackerObjectUid() ).get(), e,
-                    bundle ) ) );
-    }
-
-    private Map<TrackerDto, List<EventActionRule>> filterEvents( List<RuleEffects> ruleEffects, TrackerBundle bundle )
-    {
-        return ruleEffects.stream()
-            .filter( RuleEffects::isEvent )
-            .filter( e -> bundle.findEventByUid( e.getTrackerObjectUid() ).isPresent() )
-            .collect( Collectors.toMap( e -> bundle.findEventByUid( e.getTrackerObjectUid() ).get(),
-                e -> buildEventActionRules( bundle.findEventByUid( e.getTrackerObjectUid() ).get(), e, bundle ) ) );
-    }
-
-    private List<ActionRule> buildEnrollmentActionRules( Enrollment enrollment, RuleEffects ruleEffects,
-        TrackerBundle bundle )
-    {
-        List<Attribute> payloadTeiAttributes = bundle.findTrackedEntityByUid( enrollment.getTrackedEntity() )
-            .map( TrackedEntity::getAttributes )
-            .orElse( Collections.emptyList() );
-        List<Attribute> attributes = mergeAttributes( enrollment.getAttributes(),
-            payloadTeiAttributes );
-
-        return ruleEffects
-            .getRuleEffects()
-            .stream()
-            .map( effect -> buildEnrollmentActionRule( effect.ruleId(), effect.data(),
-                effect.ruleAction(), attributes ) )
-            .filter( Objects::nonNull )
-            .collect( Collectors.toList() );
-    }
-
-    private ActionRule buildEnrollmentActionRule( String ruleId, String data, RuleAction ruleAction,
-        List<Attribute> attributes )
-    {
-        if ( ruleAction instanceof RuleActionAssign )
-        {
-            RuleActionAssign action = (RuleActionAssign) ruleAction;
-            return new AssignActionRule( ruleId, data, action.field(), attributes );
-        }
-        if ( ruleAction instanceof RuleActionSetMandatoryField )
-        {
-            RuleActionSetMandatoryField action = (RuleActionSetMandatoryField) ruleAction;
-            return new MandatoryActionRule( ruleId, action.field() );
-        }
-        if ( ruleAction instanceof RuleActionShowError )
-        {
-            RuleActionShowError action = (RuleActionShowError) ruleAction;
-            return new ErrorActionRule( ruleId, data, action.field(), action.content() );
-        }
-        if ( ruleAction instanceof RuleActionShowWarning )
-        {
-            RuleActionShowWarning action = (RuleActionShowWarning) ruleAction;
-            return new WarningActionRule( ruleId, data, action.field(), action.content() );
-        }
-        if ( ruleAction instanceof RuleActionErrorOnCompletion )
-        {
-            RuleActionErrorOnCompletion action = (RuleActionErrorOnCompletion) ruleAction;
-            return new ErrorOnCompleteActionRule( ruleId, data, action.field(), action.content() );
-        }
-        if ( ruleAction instanceof RuleActionWarningOnCompletion )
-        {
-            RuleActionWarningOnCompletion action = (RuleActionWarningOnCompletion) ruleAction;
-            return new WarningOnCompleteActionRule( ruleId, data, action.field(), action.content() );
-        }
-        if ( ruleAction instanceof RuleActionError )
-        {
-            return new SyntaxErrorActionRule( ruleId, data );
-        }
-        return null;
-    }
-
-    private List<EventActionRule> buildEventActionRules( Event event, RuleEffects ruleEffects, TrackerBundle bundle )
-    {
-        ProgramStage programStage = bundle.getPreheat().getProgramStage( event.getProgramStage() );
-        Set<DataValue> dataValues = event.getDataValues();
-
-        return ruleEffects
-            .getRuleEffects()
-            .stream()
-            .map( effect -> buildEventActionRule( effect.ruleId(), effect.data(),
-                effect.ruleAction(), dataValues ) )
-            .filter( Objects::nonNull )
-            .filter( effect -> isDataElementPartOfProgramStage( effect.getField(), programStage ) )
-            .filter( effect -> needsToValidateDataValues( event, programStage ) )
-            .collect( Collectors.toList() );
-    }
-
-    private EventActionRule buildEventActionRule( String ruleId, String data, RuleAction ruleAction,
-        Set<DataValue> dataValues )
-    {
-        if ( ruleAction instanceof RuleActionAssign )
-        {
-            RuleActionAssign action = (RuleActionAssign) ruleAction;
-            return new EventActionRule( ruleId, data, action.field(), action.content(), RuleActionType.ASSIGN,
-                dataValues );
-        }
-        if ( ruleAction instanceof RuleActionSetMandatoryField )
-        {
-            RuleActionSetMandatoryField action = (RuleActionSetMandatoryField) ruleAction;
-            return new EventActionRule( ruleId, data, action.field(), null, RuleActionType.MANDATORY_VALUE,
-                dataValues );
-        }
-        if ( ruleAction instanceof RuleActionShowError )
-        {
-            RuleActionShowError action = (RuleActionShowError) ruleAction;
-            return new EventActionRule( ruleId, data, action.field(), action.content(), RuleActionType.ERROR,
-                dataValues );
-        }
-        if ( ruleAction instanceof RuleActionShowWarning )
-        {
-            RuleActionShowWarning action = (RuleActionShowWarning) ruleAction;
-            return new EventActionRule( ruleId, data, action.field(), action.content(), RuleActionType.WARNING,
-                dataValues );
-        }
-        if ( ruleAction instanceof RuleActionErrorOnCompletion )
-        {
-            RuleActionErrorOnCompletion action = (RuleActionErrorOnCompletion) ruleAction;
-            return new EventActionRule( ruleId, data, action.field(), action.content(),
-                RuleActionType.ERROR_ON_COMPLETE, dataValues );
-        }
-        if ( ruleAction instanceof RuleActionWarningOnCompletion )
-        {
-            RuleActionWarningOnCompletion action = (RuleActionWarningOnCompletion) ruleAction;
-            return new EventActionRule( ruleId, data, action.field(), action.content(),
-                RuleActionType.WARNING_ON_COMPLETE, dataValues );
-        }
-        if ( ruleAction instanceof RuleActionError )
-        {
-            return new EventActionRule( ruleId, data, null, null, RuleActionType.SYNTAX_ERROR, dataValues );
-        }
-        return null;
-    }
-
-    private List<Attribute> mergeAttributes( List<Attribute> enrollmentAttributes, List<Attribute> attributes )
-    {
-
-        List<Attribute> mergedAttributes = Lists.newArrayList();
-        mergedAttributes.addAll( attributes );
-        mergedAttributes.addAll( enrollmentAttributes );
-        return mergedAttributes;
-    }
-
-    private boolean isDataElementPartOfProgramStage( String dataElementUid, ProgramStage programStage )
-    {
-        if ( StringUtils.isEmpty( dataElementUid ) )
-        {
-            return true;
-        }
-
-        return programStage.getDataElements()
-            .stream()
-            .map( BaseIdentifiableObject::getUid )
-            .anyMatch( de -> de.equals( dataElementUid ) );
+        bundle.setEnrollmentActionRules( actionRuleBuilder.buildEnrollmentActionRules( ruleEffects, bundle ) );
+        bundle.setEventActionRules( actionRuleBuilder.buildEventActionRules( ruleEffects, bundle ) );
     }
 
     private List<RuleEffects> calculateEnrollmentRuleEffects( TrackerBundle bundle, TrackerPreheat preheat )
